Further cleanup of timeline-data
