+++ core/src/main/resources/org/jobrunr/dashboard/frontend/src/components/utils/state-names.js
@@ -6,6 +6,8 @@ export const SUCCEEDED = "SUCCEEDED";
 export const FAILED = "FAILED";
 export const DELETED = "DELETED";
 
+export const RUN_STEP_ONCE = "RUN_STEP_ONCE";
+
 export const STATE_LABELS = {
     AWAITING: "Pending",
     SCHEDULED: "Scheduled",
@@ -14,6 +16,7 @@ export const STATE_LABELS = {
     SUCCEEDED: "Succeeded",
     FAILED: "Failed",
     DELETED: "Deleted",
+    RUN_STEP_ONCE: 'Step (runStepOnce)',
 };
 
 export const END_STATES = [SUCCEEDED, FAILED]
